+++ node/middlewares/applyManualPrices.ts
@@ -1,7 +1,19 @@
-/* eslint-disable @typescript-eslint/ban-ts-comment, @typescript-eslint/no-unused-vars */
-// @ts-ignore
+import { applyApportionment } from '../utils'
+
 const applyManualPrices = async (ctx: Context, next: () => Promise<any>) => {
-  // TODO: Add manual prices to orderForm and identifications to Custom Data
+  const apportionedPayload = applyApportionment(
+    ctx.state.parsedRequestProtocol,
+    ctx.state.externalProviderResponse
+  )
+
+  ctx.clients.checkout.updateItems(
+    ctx.state.orderFormId as string,
+    apportionedPayload,
+    'AUTH_TOKEN'
+  )
+
+  ctx.status = 204
+
   await next()
 }