+++ node/utils/index.ts
@@ -1,4 +1,5 @@
 import { parseOrderFormToProtocol } from './parseOrderFormToProtocol'
 import { getOrderFormId } from './getOrderFormId'
+import { applyApportionment } from './applyApportionment'
 
-export { parseOrderFormToProtocol, getOrderFormId }
+export { parseOrderFormToProtocol, getOrderFormId, applyApportionment }